Committing an (attempted) conversion to float64 from Period to patch up the actual functionality for now...

+++ up.go
@@ -26,7 +26,7 @@ type Duration struct {
 func main() {
     // TODO: deal with errors
     f, _ := GetRawUptime()
-    d := NewDuration(f)
+    d := NewDuration(float64(f))
 
     // TODO: offload into more flexible output function(s)
     fmt.Printf("%dd⋅%dh⋅%dm", d.Days + d.Weeks * 7, d.Hours, d.Minutes)
